Fix load-import --local with proxy
When load-import is called with --local the load is imported locally
from the active controller. When it's called with "--os-region-name
SystemController", the request is intercepted by the dc-api-proxy.

This commit fixes a bug that happens when the request is intercepted
by the dc-api-proxy. When using --local, the function receives the
params through "request.body" and when not using it the params come
through request.POST. Only the second way was being treated.

Test plan:
  PASS: load-import with --os-region-name SystemController,
  --inactive and --local should be completed if using an upgradable
  load
  PASS: load-import with --os-region-name SystemController,
  --inactive and --local should fail if using a non-upgradable load
  PASS: load-import with --os-region-name SystemController, --active,
  and --local should complete using a load from the same version
  PASS: load-import with --os-region-name SystemController and
  --inactive should be completed if using an upgradable load
  PASS: load-import with --os-region-name SystemController and
  --inactive should fail if using a non-upgradable load

Closes-bug: 2043511

Change-Id: Id5c02fdc35c43b4134fd34592de36d35f9fc88b6
Signed-off-by: Lindley Werner <Lindley.Vieira@windriver.com>