Improve the validation of load deletion in the upgrade strategy
When the upgrade strategy finishes, the last step is to delete the
existing load, although the load will be deleted, the deletion
validation will fail because it will not be able to get the keystone
client of the given subcloud.

This occurs because the logic to check the delete validation is
trying to get the keystone client from the subcloud name, instead of
its region.

This commit fixes the mechanism providing the region name when trying
to get the keystone client.

Test Plan:
PASS: Finish System controller upgrade, then perform a subcloud rename
      dcmanager subcloud unmanage <subcloud-name>
      dcmanager subcloud update subcloud1 --name <new_name>
      Create an Upgrade strategy
      Apply upgrade strategy

Story: 2010788
Task: 49094

Change-Id: I0e5c18891715df44285372a34eb8a7c75dc037b1
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>