Add timeout to restore playbook
In this commit, a 1h timeout was add to the restore playbook
to prevent the playbook to hang indefinitely.

Test Case:
PASS: Create a backup and restore a subcloud successfully.
PASS: Create a backup and add a pause to the restore playbook,
run the restore and verify that the timeout was triggered.

Closes-bug: 2045172

Change-Id: Ica53b87283dd26c6252c354692b146228248ac2b
Signed-off-by: Christopher de Oliveira Souza <Christopher.DeOliveiraSouza@windriver.com>