Mute RVMC info logs on subcloud deploy abort of install
Normal RVMC logs go to Ansible playbook log as that is the process
that called it. In the case of subcloud deploy abort, RVMC is used to
shutdown the subcloud. Since the process calling RVMC is dcmanager,
the logs go to /var/log/dcmanager/dcmanager.log, generating
lots of unnecessary info level logs.

Test Plan:

PASS: Verify that no info level RVMC logs are being generated in
      /var/log/dcmanager/dcmanager.log while running the command
      'subcloud deploy abort'.
PASS: Verify that there is no impact to RVMC logs captured in
      Ansible playbook log file during installation.

Story: 2010144
Task: 49233

Change-Id: I447b69d385bbc65c74a97fa8be482e067c1cddd3
Signed-off-by: lzhu1 <li.zhu@windriver.com>