Fix 'secondary' and 'rehome-pending' subclouds stuck at 'online'
This commit makes 'rehome-pending' subclouds auditable while they are
still online (endpoint audits are still skipped, as the subcloud is
unmanaged). Secondary subclouds are still not audited but their
availability status will be automatically set to 'offline' when its
deploy status is set to 'secondary'.

In the original design, 'secondary' and 'rehome-pending' subclouds are
not supposed to be audited, this creates the issue where the subclouds
get stuck with the 'online' availability status, preventing the user
from being able to delete it.

This commit also fixes an issue where it was not possible to set the
endpoint status to 'unknown' for 'secondary' subclouds.

Test Plan:
1. PASS - Run 'dcmanager subcloud unmanage --migrate' for an online
          subcloud and verify that:
          - All endpoint statuses were set to 'unknown';
          - Subcloud was still audited, but each endpoint audit was
            skipped;
          - After turning off the subcloud, its availability-status
            changed to 'offline' and the audits started being skipped.
2. PASS - Manage back the rehome-pending subcloud, verifying that:
          - It initially becomes 'managed' while still 'offline';
          - Audit starts running again, eventually setting the
            availability-status back to 'online';
          - Endpoint statuses started becoming 'in-sync' again.
3. PASS - Set the subcloud to 'rehome-pending', and then set it to
          'secondary', verify that the subcloud becomes 'offline' and
          that audits are skipped (all endpoint status should be set to
          'unknown').

Closes-Bug: 2047439

Change-Id: Ia21faf469aacee6f70e5b4fe6471b019ae057e13
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>