Fix resume from create passing new bootstrap-values
When resuming the subcloud deployment, a list of deploy states to run
is built based on the current deploy status and available deploy
files.
When resuming a manually installed subcloud from create-complete
providing new bootstrap_values, an error occurred because the list
of deploy states to run was being modified during the loop.
This commit ensures this list is not modified in the loop, but rather
creates a new list appending the states to be executed.

Test plan:
  - PASS: Resume a manually installed subcloud (no install values on
          db) providing a new bootstrap_values file and verify the
          deployment was correctly resumed and the subcloud was
          bootstrap/configured to completion.

Story: 2010756
Task: 49362

Change-Id: I6a6d25e0a4cbcca4a3cd3c57219285eb24c3f00f
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>