Increase the number of audit worker threads
In this commit, the number of audit worker threads was increased
from 100 to 150 in order to prevent the audit to fall behind.
If the audit doesn't have enough threads to handle all the audit
requests, it will cause a queue of audit requests waiting for a free
thread to start auditing a subcloud. If this behavior continues,
it will increase the size of the queue and the audit will fall behind,
taking hours to empty the queue and stabilize again.

Test Plan:
PASS: Leave the audit running for one day and verify that the audit
didn't fall behind.
PASS: Upload and Apply a patch to systemcontroller and verify
that the subclouds went to out-of-sync.

Closes-bug: 2047852

Change-Id: I4e129b56ca0abfaea1e9998e466a439c76b33187
Signed-off-by: Christopher de Oliveira Souza <Christopher.DeOliveiraSouza@windriver.com>