Fix subcloud abort while on unabortable task
A bug on run_abort function caused all abort operations to continue
even if the process was killed by an external source, like playbook
failure, while waiting for the unabortable flag to be cleared.
This commit changes the behaviour to return True only if the process
was killed by the abort operation. A further improvement was made to
the unabortable flag check, stopping the loop as soon as the process
is killed by an external source, instead of waiting 10 minutes like
before.

Test plan:
  - PASS: Abort the bootstrap during an unabortable task. Manually
          kill the ansible process and verify the deploy status
          is 'bootstrap-failed' and the log message indicating
          the abort won't continue is shown in dcmanager.log.
  - PASS: Abort the bootstrap during a normal task. Verify the abort
          succeeded and the deploy status is 'bootstrap-aborted'.
  - PASS: Abort the bootstrap during an unabortable flag without
          forcing any issues. Verify the bootstrap was aborted
          and the deploy status is 'bootstrap-aborted'.

Story: 2010756
Task: 49362

Change-Id: I31af827d15ee8953b8e25dd438cc190de87d3b15
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>