Remove the require_audit_flag while checking remote SPG status
This commit removes the require_audit_flag from the peer group
audit conditional, preventing local subclouds from being
incorrectly set to secondary status when a non-primary site
comes online.

Additionally, it removes the "subcloud migrate" API, as the
"subcloud-peer-group migrate" can be re-executed to migrate a
single subcloud that failed to migrate in a previous attempt.

Test Plan:
PASS - Power off the primary site and migrate the subclouds to
       the non-primary site. Power on the primary site after the
       migration is complete, checking the subclouds' status on
       both sites. On the primary site, the subclouds should be
       secondary, and on the non-primary site, they should be
       managed and online.
PASS - With both the primary site and non-primary site online,
       migrate the subclouds away and back, verifying that
       the operation completes successfully.

Closes-Bug: 2049336
Depends-On: I793bee7d680d59a4d29a585c88c8c1689976ee1b

Change-Id: Ib702fcf318d01bd7bfa8a19ec2aee516f6089cb0
Signed-off-by: Zhang Rong(Jon) <rong.zhang@windriver.com>