Restrict subcloud deploy configure to primary site
Restrict subcloud's secondary sites (the peer site where
subcloud in SPG can be migrated to) from reconfiguring
the subcloud. This is necessary because subcloud reconfig
may update the OpenStack endpoints/routes/network IPs and impact
the ability to migrate the subcloud back to the primary site.

- Introduced a check in "deploy configure" to ensure that
the target subcloud is managed by its primary site. The site
status is inferred from the subcloud's peer group priority on the site.

Test Plan:
Setup a DC system with GR configuration (2 sites + subclouds) and
verify the  "subcloud deploy config" command with the following arrangements:
1. PASS: For a subcloud that's not part of protection group
         (not in any SPG / not part of GR),
         ensure that the operation is not restricted.
2. PASS: Add subcloud to SPG and keep it managed by primary site,
         ensure that the operation is not restricted.
3. PASS: Add system peer, create SPG association with a secondary
         site, and migrate the subcloud from #2. Verify that "deploy
         config" is restricted and it fails with following error
         message: "Subcloud must be managed by its primary site"
4. PASS: Using the subcloud from #3, migrate it back to the primary
         site. Attempt to deploy config and ensure that the operation
         is not restricted.

Closes-Bug: 2054161

Change-Id: I7e986771a32882bfad917c85439e6f6b99dfe173
Signed-off-by: Salman Rana <salman.rana@windriver.com>