Temporarily set software audit to always be in-sync
Since USM is still a work-in-progress feature, errors could occur,
causing the software audit to be out-of-sync, which also results in
the subcloud being out of sync.

This commit changes the behavior of the software audit to always report as in-sync, even if issues are found with the software service and the USM endpoints.

Known issues:
- Story = 2010676
- Task = 49666: Add support for https in USM software client
- Task = 49668: Copy GA metadata file to USM location in the Subcloud

Test Plan:
PASS - Manage a subcloud and check if the software audit is
in-sync even if the software service is disabled.
- Other audits in-sync.
- Subcloud in-sync.

Story: 2010676
Task: 49598

Change-Id: I596219f1fa4c42e6be7984f06618349c54f9448c
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>