Log subcloud health output upon failed health check
Log subcloud health output for quick diagnosis. With this change,
the user would not need to log into the subcloud to check
health output. Sometimes, the health condition is resolved by
the time the user runs the system health-query command in the
subcloud. As the result, more time would be required to determine
what caused the failed health check in the first place.

Test Plan:
  - Verify successful subcloud backup with the change.
  - Induce a health condition (e.g. management affecting
    alarm). Verify that subcloud backup request is rejected
    and the subcloud health output is captured in
    dcmanager-api.log

Closes-Bug: 2056721

Change-Id: I32fea354f9cf594ea45d412359a9090e7b1bfb83
Signed-off-by: Tee Ngo <tee.ngo@windriver.com>