Fix bootstrap_address parameter name during PGA sync
This commit changes the bootstrap address parameter from
"bootstrap-address" to "bootstrap_address" during the subcloud update
call made during the PGA sync operation. This fixes the issue where
the bootstrap_address was not being updated on the peer site, as the
subcloud update API expects the "bootstrap_address" parameter, with
an underscore.

Test Plan:
1. PASS - Create a peer group association and let it do the initial
          sync. Modify the bootstrap-address of the subcloud using
          the subcloud update command and then run the PGA sync
          command. Verify that the rehome_data of the secondary
          subcloud was updated with the new address.

Closes-Bug: 2057973

Change-Id: Ib5786a56c90f771b940e740bc095ebc8168d2830
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>