Add 'usm' to subcloud users to replicate
Add usm to the list of users whose credentials needs to be
replicated in the subcloud. For any software commands to work
in side subcloud, after it is made to 'managed' state, the 'usm'
user credentials needs to be replicated in the subcloud.

Test Plan:
PASS: Install DC subcloud, ensure it is in managed state,
      and execute software commands (Eg. software list)

Closes-Bug: 2063460

Change-Id: I4af841dcc51dc7fea2a6a12a37728cb9e0f8b59c
Signed-off-by: Joseph Vazhappilly <joseph.vazhappillypaily@windriver.com>