Create seed iso for subcloud enrollment
Implement cloud-init nocloud seed iso generation, required to
reconfigure and prepare factory-installed nodes for the
enrollment process. The seed ISO contains the networking and user data,
it specifies the configuration applied upon booting the standalone system,
done as the initial step of the subcloud enrollment process.

Overall, this commit introduces the SubcloudEnrollment class with the
initial enrollment operations to:
   - Build the data files required
     to update the user and OAM network configuration
   - Generate the seed iso with the data files
   - Serve the iso for remote use

These will be utilized in subsequent changes where the
generated iso will be mounted to a factory installed node using RVMC.

Furthermore, these changes will be revisited when integrating with
the enrollment APIs to ensure that the payload is interpreted
correctly as iso_values from the provided
install + bootstrap values.

Test Plan:
1. PASS: Validate seed iso generation:
         Invoke generate_seed_iso with payload
         and ensure seed.iso is created at
         /opt/platform/iso/<rel-version>/nodes/<subcloud>
         Furthermore, ensure the iso is available and served
         by lighttpd at:
         /var/www/pages/iso/<rel-version>/nodes/<subcloud>

2. PASS: Validate iso regeneration:
         ensure previous seed iso is
         cleaned up and iso is regenerated.

3. PASS: Verify temp dirs are cleaned up

4. PASS: Validate contents of seed iso:
         Mount seed iso and ensure both
         network-config and cloud-config exists
         and they are correctly generated
         based on the payload.

Story: 2011100
Task: 50053

Change-Id: Icef8258852746aef2d0a3f025a1fe85fff93980e
Signed-off-by: Salman Rana <salman.rana@windriver.com>