Block dcmanager subcloud delete execution during active operation
Add a validation in dcmanager subcloud delete command to
ensure that a subcloud with an active operation, such
as bootstrapping, configuring and installating can not
be deleted and, instead, dcmanager subcloud abort
should be executed.

Test plan:
1. Run the delete command for subclouds in bootstrapping
and configuring deploy status and verify that an error
is raised stating that the operation is invalid.
2. Run the delete command for subclouds in bootstrap-failed
and complete states and verify that the subcloud is deleted.

Closes-Bug: 2065595

Change-Id: Ic9dbe0ddce11cc211b64b0ba687b012594b88351
Signed-off-by: Raphael Lima <Raphael.Lima@windriver.com>