Enable dc orchestration VIM strategy apply for sw-deploy
This commit enables the apply_vim_software_strategy state
for USM type of orchestration. The sw-deploy strategy will
then be applied and VIM on the subcloud will in turn carry
out a series of software deploy actions on the subcloud.

Test Plan (in-service and reboot required patches):
PASS: Execute the sw-deploy-strategy create/apply for subcloud.
- Upload/Apply/Install the release on the system controller.
- Upload the release on the subcloud.
- Manage the subcloud.
  - The subcloud should be out-of-sync for USM.
- Create the sw-deploy-strategy <release-id>.
- Apply the sw-deploy-strategy.
- The subcloud should have the release in the deployed state.
  - The subcloud should be in-sync for USM.

Story: 2010676
Task: 50199

Change-Id: Id6bf06e38323b42f4a5e6e81b9c6237d315bf918
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>