Add file validation to subcloud deploy resume
Adds a file validation to subcloud deploy resume command in order to
accept payloads that contain only the required data files, quitting the
execution if any other one is sent, e.g. a subcloud in
bootstrap-complete state should only receive the deploy config file
since that's the remaining step to execute. If a bootstrap values or
install values is sent, the execution is terminated.

Test plan:
1. PASS: In a bootstrapped subcloud, execute the resume command sending
both the bootstrap values and install values, resulting in a failure.
2. PASS: In a bootstrapped subcloud, execute the resume command sending
only the bootstrap values, resulting in a failure.
3. PASS: In a bootstrapped subcloud, execute the resume command sending
only the install values, resulting in a failure.
4. PASS: In a bootstrapped subcloud, execute the resume command without
sending any files, resulting in the process starting.
5. PASS: In a bootstrapped subcloud, execute the resume command sending
only the deploy config file, resulting in the process starting.

Closes-Bug: 2067326

Change-Id: I6221ea6c15e119d7451611309bff557a6e2b783c
Signed-off-by: Raphael Lima <Raphael.Lima@windriver.com>