Prestaging support for --for-sw-deploy/--for-install
Add USM prestaging support. This commit introduces support for the
--for-sw-deploy and --for-install prestaging options. The --for-install
option is the default, and is the equivalent of the previous release
prestaging behaviour - the subcloud ostree repo is prestaged to the
platform-backup partition. The --for-sw-deploy option is the new
prestaging behaviour for USM major and minor releases. This commit
focuses on prestaging minor release updates.

The --for-sw-deploy option requires synchronization of ostree commits
and /opt/software/metadata contents from the system controller to the
subcloud.

Notes on --release YY.MM and --release YY.MM.nn handling:
- The --release value can either be a major release, in format YY.MM, or
  a specific minor release, in format YY.MM.nn.
- If format is YY.MM, the behaviour is to prestage ALL minor releases
  present on the system controller to the subcloud.
- If format is YY.MM.nn, the behaviour is to prestage ONLY the given
  minor release
    - Note: there is no sanity checking for dependent minor updates.
      If update YY.MM.nn is given but YY.MM.nn-1 is not present, the
      operation is currently allowed. This will be addressed in a future
      commit (see TODO below)
- TODO: Future commit will expand on this behaviour for YY.MM.nn to
  prestage all missing commits up to and including the YY.MM.nn commit

This update also fixes the /opt/platform/iso/YY.MM/ostree_repo bind
mount if it has not already been established via a subcloud install.

Test Cases
PASS
- For prestaging operation, initiate ansible playbook with correct
  values of for_install / for_sw_deploy variables, and correct
  software_version / software_major_release values corresponding to the
  values input by the client.
- Verify that a stale or missing /opt/platform/iso/YY.MM/ostree_repo
  bind mount is re-established upon start of prestage operation.

Story: 2010676
Task: 50322

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I5cc08499e44bcb66eecdb28dc227cc032ee9890a